Update Math.cpp
I changed the namespace name from math to Math. It is generally a good practice to use uppercase letters for the first letter of namespaces, classes, and constants.
I changed the function names to follow the CamelCase naming convention, where each word in the name is capitalized except for the first word. This is a common naming convention in C++.
I added the std:: prefix to the sqrt and pow functions to specify that they are part of the std namespace. This is generally a good practice to avoid naming conflicts with other functions that might have the same name.